cleanup 1 of GitP08;
made an interface to ML, made a wrapper for Rack requests, got it
working for both rack/sinatra tests and also from an IE browser.
Tightened up the code. Still unhappy I don't know how to create a Rack
request on demand for my other adaptors, but this is moving me forward.